requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
Automat==22.10.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
build==1.0.3
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.2.0
cryptography==42.0.4
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
et-xmlfile==1.1.0
executing==2.0.1
ffmpy==0.3.1
filelock==3.13.1
FLAML==2.1.1
FlashRank==0.1.66
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.48.1
frozenlist==1.4.1
googleapis-common-protos==1.62.0
googlesearch-python==1.2.3

h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.2
jedi==0.19.1
Jinja2==3.1.3
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.22
langchain-openai==0.0.6
nltk==3.8.1
numpy==1.26.4
openai==1.12.0
python-dotenv==1.0.1
qdrant-client==1.7.3
tiktoken==0.6.0
Werkzeug==3.0.1
rank-bm25==0.2.2